Handle instances of subclasses of core classes when wrapping objects in the pg_json extension (Fixes #1631)
This switches to a case-based approach instead of a hash-based
approach, which handles subclasses automatically.  Due to the small
number of classes needing wrapping, the case-based approach may
even be faster than the hash-based approach in the common case.

Deprecate the hash constants previously used to find the wrapper
classes.