Rename plural_association_methods plugin to association_multi_add_remove
Plural describes the method name, but doesn't really describe the
purpose of the plugin.

Use a transaction when adding/removing multiple associated objects,
so if one of the later ones fail, you don't end up in a state where
only part of the changes are saved in the database.  Handle sharding
correctly when creating the transaction.

Add a check that the method name you are defining for adding/removing
multiple associations is not the same as the one for adding/removing
a single association (it would be by default for uncountable nouns).
Also, allow customizing the method names defined using the
:multi_add_method and :multi_remove_method association options.