Minor fixes to association_multi_add_remove setter
Do not define a setter method if we didn't define both a multi_add
and multi_remove due to name conflicts.

Don't try to readd objects that are already associated. In the
many to many cases, that can cause duplicate issues in the join
table if there is no unique constraint, and unique constraint
violations if there is a unique constraint. It causes unnecessary
queries in the one to many case.

We can return nil from the setter.  Assignment in Ruby always
returns the LHS, regardless of what the method returns, and I
don't think the return value matters if called as a method.