Make dirty plugin work correctly with typecast_on_load plugin
Introduce Model#_clear_changed_columns private method, which is
called with the reason we are clearing the columns.  If the
reason is :initialize, it is because the model was just initialized
and none of the columns should be considered as changed.  In this
case, the dirty plugin can clear the initial column values.