Warn when calling Sequel::JDBC::Postgres::Dataset#with_fetch_size (Fixes #1665)
The JDBC PostgreSQL driver ignores the fetch size unless autocommit
is turned off.  Sequel doesn't modify the autocommit setting, so
the default is probably on.

A better fix would be to implement use_cursor in the jdbc/postgresql
adapter and have with_fetch_size use that, similar to what the
postgres adapter does.