Add association_lazy_eager_option plugin for supporting :eager option for association method
This plugin makes it easier to eagerly load nested associations
when lazy loading associations.  That was already possible by
passing a block to the association method and calling eager inside
of it.  However, passing a block ignores any cached association.

The main advantage of this plugin is that it does not ignore
cached associations, it just returns the cached association without
doing any loading.  This is useful in case you have code where the
association and the nested associations are both already cached.