Add :generated schema entry on PostgreSQL 12+ and SQLite 3.31+ for whether the columns is generated
MySQL already set this schema entry, so follow the same format.

SQLite was previously not including generated columns in the schema,
since it treated them like hidden columns.  Start using table_xinfo
pragma on SQLite 3.31+ so that hidden columns are show, but still
filter out columns that are not generated.  Guess whether they are
generated by looking for GENERATED ALWAYS at the end of the type,
and if so, remove that from the type.