Dump generated columns as generated columns when using the schema_dumper with :same_db option on PostgreSQL 12+ (Fixes #1680)
This ignores defaults for generated columns by default.  For
non-same_db dumps, this shouldn't have much effect, as MySQL
and SQLite do not include the default values in the parsed
schema, and the parsed value on PostgreSQL is unlikely to
have a valid ruby default.

For same_db dumps, this fixes handling of generated columns
on PostgreSQL.  Previously, the default value was treated
as a regular column default instead of as a generated column
value.  Now, instead of using the :default key, the
:generated_always_as key is used, which fixes such cases.