Rename skip_saving_generated_columns and skip_saving_columns and modify it (Fixes #1681)
This makes the plugin more general in multiple ways.  It allows
customizing the columns to be skipped, and it handles the
new model instance creation case as well as the case when
save_changes is used.

In order to handle save_changes, make some minor changes to
Model#_save to extract a _save_update_changed_colums_hash, similar
to the existing _save_update_all_columns_hash.  This also correctly
calls _clear_changed_columns when using save_changes.