Fix method visibility issues in model, plugin, extension, and adapter code
This mostly fixes cases where private methods have been made public
by accident in an included module.

There is one case where this changes the base visibility, and that
is Model#_insert_values.  This was intended to be a private method,
but defining an alias method takes the visibility from the original
method, which in this case was public.

This uses the visibility_checker gem to do the checking, and
adds spec_*_vis tasks to run the specs with visibility checking.