Add fiber_concurrency extension, for using Fiber.current instead of Thread.current for checking out connections
This allows fibers of the same thread to use different connections.

Given the following setup:

  ds = DB.select(1).union(DB.select 2)

This extension can handle cases such as implicit fiber usage in
enumerators:

  enums = 4.times.map{ds.to_enum}
  2.times do
    enums.each{|e| p e.next}
  end
  enums.each do |e|
    begin
      p e.next
    rescue StopIteration
      # Need to do this so the iteration ends
    end
  end

or direct use of fibers:

  fibers = 4.times.map{Fiber.new{ds.each{|r| Fiber.yield [Fiber.current, r]}}}
  # One additional time so that the Dataset#each block finishes
  3.times do
    fibers.each do |f|
      p f.resume
    end
  end

With Sequel's defaults, these result in a DatabaseError as the same
connection is used by multiple fibers simultaneously, if the adapter
is streaming results.  With the fiber_concurrency extension, these
examples work as they use separate connections.

There is a tradeoff here.  If you try with the number of fibers
greater than the number of allowed connections, you'll get a pool
timeout because one fiber will be waiting to checkout a connection
and the other fibers will not be running or able to check a
connection in (a type of deadlock).  There is work being done in
Ruby to allow for automatic fiber switching when trying to lock an
already locked mutex, at least for certain kinds of fibers.