Make Sequel::Model.plugin issue deprecation warning if loading plugin with arguments and block if plugin does not accept arguments/block
If the plugin doesn't have a configure or apply method, then the
arguments and block passed to Model.plugin would be ignored. This
will cause problems if a block or optional arguments are added
later.  It's better to raise in this case, but in Sequel 5 we'll
just issue a deprecation warning.