Make Model.finalize_associations not affect instance-specific associations
Previously, Model.finalize_associations could break associations
that used instance-specific code in the association block, and
possibly other options (such as :dataset).  The approach taken here
is a tad heavy-handed, by disabling the finalize optimization
completely for instance-specific datasets.  However, the percentage
of associations that are actually instance specific is quite low, so
I don't think this matters.  This does provide a performance reason
to set instance_specific: false for associations with blocks that are
not instance-specific, though.