Copy composite unique constraints when emulating alter table operations on SQLite (Fixes #1704)
Don't copy them if the alter table operation is dropping a unique
constraint.  So this won't handle a case where there are multiple
unique constraints and only one is being dropped.

Make sure not to add a duplicate composite unique constraint if it
would be the same as the composite primary key constraint.