Always skip association placeholder loader if :dataset option given
The placeholder loader does not consider the :dataset option, so it
should never be used even if the association is marked as not
instance specific.  In general, a :dataset option that isn't
instance specific is unlikely, as otherwise you would get the same
associated values for every instance of the model.