Make Model.finalize_associations not break with instance-specific associations
The previous approach did fix the issue, but it also disabled the
optimization.  With this approach, we keep the optimization except
for the two methods that break for instance specific associations.

This could potentially affects external association types that
have finalize_settings not include associated_eager_dataset and
filter_by_associations_conditions_dataset, but it seems unlikely
that anyone is doing that.