Document issue with newer jdbc-mysql gem versions
Depending on how the server is configured, you may need to
specify the serverTimezone correctly.  This seems like a bug in
the driver in the newer gem, since older gem works fine, but
it isn't an issue that can be fixed by Sequel, since Sequel
doesn't know the appropriate timezone to use.