Remove ".sql" call in cheat sheet examples
Since the cheat sheet uses the convention that the resulting SQL is
given in an adjacent comment, including the ".sql" method call in the
examples is unnecessary.  Also, sometimes the ".sql" method call was
included, and sometimes not.  Remove all occurrences of it except, of
course, in the Retrieving SQL examples.