Backout use of ActiveSupport::Duration.build
This cannot handle the difference between 30 days and 1 month.
ActiveSupport::Duration.parse doesn't work for PostgreSQL interval
types.  Even though ActiveSupport::Duration.new is considered
private API, there is no other choice but to use it.  Have it use
a hash for the parts in ActiveSupport 5.1+, and an array for the
parts in ActiveSuport < 5.1.