Add async_thread_pool Database extension for executing queries asynchronously using a thread pool
Allowing for asynchronous query execution can decrease latency in
applications that currently execute queries in serial, especially
in cases where there is a large amount of latency between the
database and the applcation.  This allows for asynchronous query
execution that returns a proxy object that will block when methods
are called on it until the results are available.

By default, if a dataset is set to asynchronous mode via
Dataset#async, methods that return results will have the related
query and method logic be executed by a thread pool.  By default,
that will always happen in a separate thread for consistency, but
in cases where it is desirable to allow preemption (so the current
thread will execute the query if it hasn't been executed yet), you
can use the :preempt_async_thread Database option.  You can use
the :num_async_threads Database option to set the number of threads
in the thread pool that will execute the asynchronous queries (this
defaults to the size of the connection pool).

Probably the worst part of async queries is that they don't support
transactions, since they run on separate database connections. So
you cannot use them with transactional testing.  There are also
some corner cases with the proxy objects when using them in a boolean
context or in case expressions.