Add the ado timestamp conversion proc to conversion_procs
This makes it so users can modify the behavior of the conversion,
instead of hardcoding it.

The previously hardcoded behavior was for picking up the current
timestamp_precision, which can be different per database, but from
my testing, Microsoft SQL Server, which uses a timestamp_precision
of 3, works with the code for a timestamp_precision of 6, due to
what nsec values it returns.

External ADO subadapters can now override the
adb_timestamp_to_application_timestamp to get custom behavior.