Test column_encryption plugin with CHECK constraints
To make unintentional data corruption of encrypted columns less
likely, it's wise to add CHECK constraints.  This tests three
types of CHECK constraints.  Two are general and should work
on all databases.  The first uses a LIKE with the 3 patterns the
data should start with (one for nonsearchable, one for
searchable, one for case-insensitive searchable).  The second
is a simple minimum length check for the base64 encoded overhead
of the encrypted column without padding.

The third CHECK constraint is checking for valid urlsafe base64,
and that's going to be database specific.  A test is added for
PostgreSQL.  This CHECK constraint is more of a failsafe, it's
very unlikely that unintentional data corruption will not be
caught by the first CHECK constraint, as the odds of a false
negative in the first check constraint are about 3/(255**3).

This also adjusts the minimum length check in decrypt to reflect
the actual minimum of (4+32+12+16+1)=65 for non-searchable columns
and 97 for searchable columns.  The previous minimums were based
on earlier development that used a 3 byte header and lacked a
padding byte in the encrypted data.