Support dropping non-composite unique constraints on SQLite (jeremyevans) (Fixes #1755)
Previously, this was only skipped for composite columns.  This is
now skipped for non-composite columns.  Note that this makes
non-composite unique constraints similar to composite unique
constraints, in that dropping any unique constraints from a table
will drop all unique constraints from the table.  As SQLite does
not support dropping constraints, and the API does not currently
support specifying which columns for the unique constraint being
dropped, there isn't a way around this.  One possible way to
handle it would be an additional option to drop_constraint that
specified the related column or columns, which could be used for
matching which autocreated unique index to skip.