Add better spec guard for column_encryption specs
This fixes cases where Ruby 2.3+ is used but the openssl extension
does not correctly support the aes-256-gcm cipher. Previously, the
code tested the cipher could be loaded, but it didn't test that the
cipher actually worked.  This tests that the cipher works correctly
when the plugin is loaded, raising a LoadError if it doesn't work
correctly.

This fixes some cases I've tested on Ruby 2.3 where the cipher
exists but does not work correctly.