Make date_arithmetic extension handle ActiveSupport::Duration values with :weeks (Fixes #1759)
Due to how date_sub works, this also requires supporting :weeks
in explicit hashes.  Since both :weeks and :days can be submitted
in a hash, handling a hash argument needs to be similar to handling
and ActiveSupport::Duration argument, so unify the handling.

To avoid having to modify all of the database-specific code, convert
weeks to days automatically.

While here, raise an exception if an invalid key is used, so we
can catch future cases where an unsupported key is used.