Disconnect connections before running the transaction test
I've seen occasional failures in these tests when the
async_thread_pool extension is used, though not in isolation. There
is probably an implicit assumption in one of the tests that the
there is only a single connection in the connection pool. Disconnecting
before running these tests should address that.