Add concurrent_eager_loading plugin, for eager loading multiple associations concurrently using separate threads
This uses the async_thread_pool Database extension to support the
concurrent eager loading of multiple associations.  Due to the
fact that eager loading modifies the associations hash of
Sequel::Model instances, and that is not thread safe, it creates
a mutex shared by all of the eagerly loaded associations, which
protects access to the objects.  The actual database queries
are executed concurrently outside of the mutex synchronization.

By default, loading this plugin has no effect, as it will only
concurrently eagerly load for datasets that have been marked
for concurrent eager loading via the eager_load_concurrently
method.  However, by using the :always option when loading the
plugin, you can make it so that all eager loads of multiple
associations use the concurrent approach.

All association types that ship with Sequel have been modified
to support this.  Users who use custom :eager_loader association
options will have to update their :eager_loader to support this.

Because a mutex may or may not be provided, to make it easier for
users with custom :eager_loader options,
Sequel.conditional_synchronize has been added.  This method either
takes a mutex or nil.  If given a mutex, it uses it to synchronize
access, otherwise it just yields to the block.

This refactors the implementation of eager_load, making it separate
the preparation of eager loading data from the execution, in order
to allow for concurrent execution.

The SEQUEL_EAGER_ASYNC environment variable can be used with the
adapter and integration specs to run the plugin enabled with the
:always option, so that all eager multiple association loads in
the specs are concurrent.