Support Dataset#returning on SQLite 3.35+
This is similar to the PostgreSQL support, which makes sense as
SQLite's RETURNING support is modeled on PostgreSQL's.

There are a couple of SQLite bugs we need to workaround to enable
this.  First is RETURNING expressions without aliases do not use
the same column names as SELECT expressions without aliases.
SELECT expressions without aliases will omit the identifier
quotes, but RETURNING expressions will not. Work around this by
automatically aliasing identifiers in RETURNING expressions.

The second issue is that using RETURNING with prepared statements
causes transactions to break with the following error:

  SQLite3::BusyException: cannot commit transaction - SQL statements in progress: COMMIT

Workaround this by having the prepared statements plugin not use
insert_select prepared statements on SQLite.