Make the unused_associations plugin track access to association reflections to determine whether associations are used
One of the main issues in the original version of the plugin was
things like association_join and similar methods access the
association metadata without calling association methods, which
shows that the associations are unused even if they are actually
used.

Have the unused_associations plugin wrap association_reflection
to track which associations are accessed through it, so those
associations can be marked as used.

This may cause false positives in cases where code iterates
over the result of associations and calls association_reflection
with each.  Such code should probably switch to using
association_reflections directly.

While here, make errors when testing the unused_associations
plugin show full backtraces from the underlying error (which
runs in a different process).