Make auto_validations plugin automatically setup no_null_byte validations
This makes auto_validations setup no_null_byte validations for
all string columns.  Many if not most databases do not allow null
bytes in strings, and most strings are not going to want to accept
null byte input. It's better to raise such errors during validation
instead of when the INSERT/UPDATE query is run.

This could potentionally cause issues in cases where the database
allows null bytes in string columns, and the user is actively using
that feature.  In such cases, you can use:

  skip_auto_validations(:no_null_byte)

inside the model to not perform such validations.