Add pg_multirange extension with support for PostgreSQL 14+ multirange types
Multirange types are similar to an array of ranges, where a value is in the
multirange if it is any of the ranges contained in the multirange.  They
is useful whenever you need check against multiple ranges that do not overlap.

I originally considered making this part of the pg_range extension, but the
behavior is different enough to justify a separate extension.  pg_multirange
is kind of a mix between pg_array and pg_range in terms of implementation and
behavior.

Because PostgreSQL range and multirange type support almost the same functions
and operators, make pg_range_ops handle both.  It needs almost no changes to
do so.  Add 3 functions to pg_range_ops that were added in PostgreSQL 14 that
related to converting between ranges and multiranges.

Update documentation for pg_range and fix typos in pg_range specs while here.