Call `super` in `#log_connection_yield` in sql_log_normalizer without loggers
This change allows using the sql_log_normalizer extension with the
sequel-activerecord_connection gem, which overrides
`#log_connection_yield` to send SQL queries to Active Record's
instrumenter, and where it's not common to have any loggers assigned to
Sequel itself. Calling `super` in this case ensures that
`#log_connection_yield` of the activerecord_connection extension is
still called.