Make association closing handle cases where clone association sets different :class option than cloned association
This is due to the moving of :class to :class_name if :class is
provided as a symbol or string.  To fix it, if :class is provided
in clone association, but :class_name is not, remove :class_name
from combined options, so :class in cloned association doesn't
override :class in clone association.