Make mysql adapter support ruby-mysql 3 API (Fixes #1795)
ruby-mysql (the pure-ruby MySQL gem) changed it's API in version 3,
so check if Mysql.init is defined, and if not, switch to a new
code path for most of the connection setup.  Add a couple aliases
to the singleton class of the connection so that the new API is
more compatible with the previous API, so fewer changes are needed
to the adapter.