Make :ruby_default schema entry for type: :datetime respect Sequel.datetime_class
If Sequel.datetime_class is Time (the default), it makes more
sense for the default value to be an instance of Time, not
DateTime.

This is a small breakage of backwards compatibility, but it's
unlikely it will effect anyone.  The values are should be the
same except for possible timezone issues, and the :ruby_default
value is not used internally except by the defaults_setter
and prepared_statements_safe plugin.