Use Class#subclasses if available in the subclasses plugin, instead of a custom Model.subclasses accessor
This uses the Class#subclasses method added in Ruby 3.1, if available.

This fixes the method naming for descendents and freeze_descendents.
Descendent is an adjective (e.g. descendent classes), descendant is
a noun (e.g. a class's descendants).  Keep the old names for now for
backwards compatibility, but they will probably be removed in
Sequel 6.

This is slightly backwards incompatible in that the order of classes
returned by Model.subclasses is no longer the order of creation.