Skip a couple postgres adapter specs temporarily in streaming always mode on pg 1.3.4
The new streaming support in pg 1.3 has a bug related to handling
streaming of procedure calls that do no return rows. I submitted
pull request #447 to fix it, but it affects these two specs. It's
very unlikely to hit this in production, because both of the
following need to be true:

* Need to be using a version sequel_pg that uses the
  pgresult_stream_any in pg 1.3.4. This version hasn't been
  released yet, but will probably be released as 1.15.0. Earlier
  versions of sequel_pg won't use this pg gem feature, and my
  assumption is pg 1.3.5 will fix the issue.

* Need to be streaming results for all queries.  This is generally
  a bad idea, and is only really available so it can be enabled
  during testing to catch bugs like this one.

* Need to be using a PostgreSQL procedure call that returns no
  results.