Support Dataset#merge* on Oracle
Oracle's MERGE support is non-standard.  In addition to non-standard
syntax, it doesn't support multiple clauses of the same type, the
DELETE clause requires the UPDATE clause, and the DELETE clause only
affects rows updated by the UPDATE clause.  T

It's incompatible enough from the standard that this skips the general
integration tests for MERGE support, and uses modified version of
the tests that checks for Oracle semantics.