Model code that does not swallow connection errors will now also not swallow disconnect errors
Before this, if the query to get the schema resulted in a disconnect
error, Sequel would swallow the error.  The query to get the columns
may still succeed, resulting in a model class that doesn't do
typecasting.

You cannot automatically retry in this case, and the same connection
may be used if the calling code is inside a Database#synchronize
block.  So the safest decision is to bail out.  Users that want to
avoid such disconnect errors should ensure a reliable connection to
the database, or use the connection_validator extension if that is
not possible.