Add require_valid_schema plugin for checking that model classes have schema parsed as expected
If the database supports schema parsing, and the dataset being
used uses a single identifier, you would probably expect schema
parsing to work for the model.  When schema parsing doesn't work,
at the very least the column typecasting doesn't work.  That's
expected for models not based on a single table, but not for other
models.  However, for historical reasons, Sequel has never required
that models have valid schema.

Requiring valid schema can break otherwise working applications,
so this cannot be turned on by default.  I'm not sure how all
adapters work when passing views instead of tables to
Database#schema.  If the adapter raises an error, this will
raise for models it shouldn't raise.  By adding this as a plugin,
people can start using for added assurance, assuming it works
in their application.  Maybe by Sequel 6 we will have the corner
cases addressed and could potentially turn this on by default.