Convert most respond_to? calls to equivalent defined? for better performance
It's probably suprising, but defined? is about 50% faster than
respond_to?, because it's built into the VM instead of being
implemented as a normal method.

Not all respond_to? calls can be converted, however. respond_to?
calls taking a second argument, using a symbol ending in =,
using a non-static symbol, or those where the return value is
returned to the user cannot be converted (defined? returns String
or nil instead of boolean).