Handle DateTime#to_time changes in Ruby 3.2 preview2
Starting in Ruby 3.2 preview2, DateTime#to_time will not give an exact
date mapping for dates before the date of calendar reform.  This makes
adjustments to avoid the use of DateTime#to_time when converting
database timestamps to Time values, since in this case.

To simplify things, remove the workaround for JRuby 9.2.0.0.  The
workaround was not needed in JRuby 9.2.0.1 and newer. JRuby 9.2.0.0
users who want to use PostgreSQL extended date support should update
to any later version of JRuby.

In addition to fixing issues on Ruby 3.2 preview2, this should also be
faster in cases where the code paths are taken, due to avoiding
allocation for intermediate DateTime objects.