Add Database#check_string_typecast_bytesize for checking bytesize of strings before typecasting
If the string is significantly longer than you would expect for
a column value, treat it as invalid.  There's no point in trying
to convert a 1MB string to an integer, decimal, float, date, time,
datetime, or some other values. The typecasting code is often
called directly with user-supplied input, and it's better to fail
fast in this case.

This tries to be conservative for the limits.  For example, 64 bit
integers fit in 21 bytes, and this uses a plain 100 or 1000 byte
limit.  However, there is still the possibility that this change
could break working code.

Weighing the pros and cons, I've decided to enable this support
by default.  Users who do not want to perform the checks can
use Dataset#check_string_typecast_bytesize= or the
:check_string_typecast_bytesize Database option to disable the
checks.

While here, fix a case where negative hexidecimal strings were not
parsed as hexidecimal when converted to integer, unlike positive
hexidecimal strings.