Include min_value and max_value schema entries for integer columns
This sets the schema entries to the minimum and maximum value
supported by the underlying database type. Almost all databases
support some subset of 1-byte, 2-byte, 3-byte, 4-byte, and/or 8-byte
integers, with signed and unsigned versions of each.

This adds integration tests so that on each database, the minimum
and maximum values used are confirmed to work, while one less than
the minimum and one more than the maximum raise errors.

This is unsupported on SQLite as it allow arbitrary sized integers
for integer types, and unsupported on Oracle as Oracle uses the
NUMBER type with variable precision for integers.