Set up automatic validation of minimum and maximum values for integer columns in auto_validations
This uses the min_value/max_value column schema entries provided
by the Database, and automatically sets up the appropriate validations,
so that in case an integer not supported by the underlying database
type is passed, you will get a validation error instead of a
database error when saving.

This adds integration testing of the auto_validations plugin, which
previously did not have integeration testing.  It also significantly
expands the extension tests.