Add pg_extended_integer_support extension for customizing behavior when literalizing a Ruby integer outside PostgreSQL bigint range
This adds support for literalizing Ruby integers outside of
PostgreSQL bigint range.  By default, the values are quoted,
causing PostgreSQL to treat them as an unknown type. To get
the historical behavior of not quoting the values when
literalizing, you can use a :raw strategy, which will cause
PostgreSQL to treat the values as numeric.