Add back the pg_auto_parameterize extension for automatically using bound variables when using postgres adapter with pg driver
This extension was originally added in 3.34.0, and removed in 4.0.0.
The extension was originally added as a proof of concept, but due to
the limitations it had, it wasn't suitable to most production
scenarios.

This updates the extension to address many of the limitations, and
to try to make it suitable for more production scenarios, including
the following changes:

* Handles integers used in LIMIT/ORDER
* Respects explicit CASTs
* Tries to convert column IN (int, ...) into column = ANY($) with
  an array parameter
* Uses the same parameter for the same object used more than once
  in a query
* Uses parameters when inserting multiple rows via Dataset#import
* Supports automatically parameterizing all of the PostgreSQL-specific
  types that Sequel ships support for in pg_* extensions (though some
  values of those types may not support automatic parameterization).
* Supports skipping auto parameterization for specific values.

Automatic parameterization does not work with some of the
optimizations that Sequel has added since the extension was originally
removed from Sequel, specifically the use of placeholder literalizers.
Add a Dataset#supports_placeholder_literalizer? method for whether the
dataset supports placeholder literalizers, and have it return
false when using the pg_auto_parameterize extension. Use this to
disable some optimizations that are enabled by default.