Correctly format hstore[] types in bound variables on PostgreSQL
Make bound_variable_array call bound_variable_arg for unhandled types,
and if it returns a string, do the same mangling done to strings. This
allows the removal of all bound_variable_array method overrides. This
fixes the hstore[] case, because that did not have a
bound_variable_array override. This wasn't caught because unlike the
other PostgreSQL-specific types, there were no tests previously for
using arrays of hstore in bound variables.