Cache reflection datasets in the postgres adapter to improve performance
This uses a shared dataset for each of the metadata methods, along
with the Dataset#{first,where_single_value,where_each,where_all}
methods to limit the number of internal datasets generated for the
metadata queries.

For the primary key and sequence queries, this switches from a
static SQL approach to using the Sequel API.

This change to cache datasets exposed an implicitly rescued exception
in the previous implementation when using the mock postgres adapter.
Have the mock postgres adapter explicitly rescue NoMethodError inside
schema_parse_table to handle cases where schema is being called
implicitly with fetch results that are not what the schema parsing
code expects.