Add advice for making migrations work with code reloading
When I was setting up a Roda & Sequel application with Zeitwerk, I was
pondering on how to solve syncing models after running migrations in
development. One idea was to disable schema caching globally altogether
in development, at the cost of more verbose SQL logs and performance.
Another idea was to somehow clear the schema cache when model classes
are unloaded, but I couldn't find any public methods that do that.

It's only when I looked at the roda-sequel-stack project that I
discovered the `cache_associations` config, which the docs say is
recommended to disable when using code reloading. And reading the Rails
source code gave me the idea to trigger the reload when the schema file
changes. Since it took me a while to arrive to these strategies, I
thought it would be useful to include them in the migration guide.