Recommend disabling association reflection cache when using code reloading
When using a code reloader that only reloads model classes that
changed, the association reflection cache should be disabled, to avoid
stale model classes still being referenced after they were reloaded.

This is not needed when using code reloaders like Zeitwerk, which
always reload all the code regardless of which file changed, which
includes all model classes. It's still recommended if you're reloading
models after running schema migrations, as it causes the table schema
cache to be automatically reloaded when the model is reloaded.