Add primary_key_lookup_check_values plugin for typecasting and checking primary key values during lookup
This plugin typecasts values given during primary key lookup, and
also checks that the values are in the allowed range (if the
database schema provided minimum and maximum values for the
primary key column).

If the primary key value cannot be typecasted correctly, or the
value after typecasting is outside the allowed range, the lookup
returns nil without executing a query on the database (causing
with_pk! to raise an exception).

This can change behavior.  Cases that previously would raise
a DatabaseError may return nil or raise NoMatchingRow instead.
When using a filter expression that isn't a primary key value
or a Symbol, Sequel::SQL::Expression, or Sequel::LiteralString
object, code that would previously return a record may return
nil or raise NoMatchingRow instead.