Use compare_by_identity hashes for Thread-keyed hashes in threaded connection pools
Using compare_by_identity hashes is faster, and this works fine
for threads because no two Thread objects are equal unless they
are identical.

This has some core spec fallout, but only because the specs are
checking the internals, and compare_by_identity hashes are not
considered equal to non-compare_by_identity hashes.