Add support for pool_class: :timed_queue on Ruby 3.2+, using a Queue for available connections
This uses the timeout option added to Queue#pop in Ruby 3.2 to
implement the queue.  This will probably become the default
connection pool on Ruby 3.2+ at some point in the future, but
currently usage is opt-in because the public API doesn't
exactly match, and it doesn't support the :connection_handling
option. In addition to passing connection pool specs, the
adapter tests do not show any failures with the timed_queue
connection tests on all four adapters I tested with.

Ruby 3.2 users are encouraged to try out this new connection
pool class and provide feedback.  To make testing easier, the
:pool_class Database option now supports String values in
addition to Symbols, so it can be specified as part of a
database connection URL.

There is a currently not a sharded_timed_queue connection pool,
so users who are using the sharding support should stick to the
default sharded_threaded connection pool.