Minor improvements to the tactical_eager_loading and table inheritance support
Instead of filtering for instances of the same class, filter for
instances of classes that support the same association.  This better
handles the case where an association is defined in the parent class.
Potentially it doesn't work correctly for cases where an association
with the same name is defined differently in a subclass, but that
seems like a bad idea.

This moves the filtering to the single_table_inheritance plugin
(which also handles class_table_inheritance, since it loads
single_table_inheritance).  This way users of tactical_eager_loading
without table inheritance do not have to pay the cost of the extra
check.  Additionally, it makes sure that only a single additional
array is created for the filtered objects, instead of doing reject
and select (which creates an additional unnecessary array).

To handle cases where tactical_eager_loading and the table
inheritance plugin can be loaded in either order, this uses
defined?(super) to check for existing behavior in both cases.

This updates the tests to actually check that the expected
associations are eager loaded, instead of just checking for
the expected SQL.  Additionally, it checks for cases where the
parent class defines the association, to confirm that it loads the
association for both objects in that case.