Support :graph_use_association_block association option to make eager_graph use the association block
This can make things simpler than using the other :graph_* options,
as long as the database is smart enough to optimize the subquery
correctly.

While working on this, I realized that many_through_many and
many_through_one associations did not support eager_graph
callbacks.  This support is implemented in the same place that
eager_graph caallbacks are, so fixing that issue was necessary
for :graph_use_association_block to work with such associations.