Improve the :min_value/:max_value support for decimal/numeric columns
This extends it to work on all databases Sequel supports.

In some cases, the precision/scale information is in the :db_type
schema entry, and in other cases, it is in the :column_size and
:scale schema entries.  Make the code handle both.  To try to get
this working on Oracle, make the Oracle schema support include a
:column_size schema entry.

To handle decimal/numeric columns used for integer types (where
scale is 0), switch column_schema_integer_min_max_values to
take a column schema hash instead of a db_type string.  This
is a backwards incompatible change, and external adapters will
need to update their code if they overrode the method.

PostgreSQL 15+ supports cases where scale is negative, and earlier
versions support cases where scale is greater than precision, so
try to handle both cases.